ntfsck: check backup boot sector check and set bitmap
NTFS backup boot sector located in last cluster/sector, or in the middle
of cluster. When it located in the middle, it's cluster bitmap should be
set not to allocate by others.
After it checked, set cluster bitmap, because ntfs set bitmap of last or middle
of cluster which is located backup boot sector.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>