ntfsck: check same type none resident attribute with different name
ntfsck_check_inode_non_resident() check all non-resident attributes of
inode. but it can't be distinguish same type with different attribute
name.
So, add checking for same type non-resident attribute with different
name. It can be distinguish with lowest_vcn field of ATTR_RECORD.
If lowest_vcn of attribute is not zero, it is not the first attribute of
same type and same name. So we can skip attribute with non-zero
lowest_vcn.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>