ntfsck: add checking entry's vcn in ntfsck_validate_index_blocks()
When $INDEX_ALLOCATION were allocated, but not used
(ie, $BITMAP cleared), ntfsck_validate_index_blocks()
skip $INDEX_ALLOCATION block which is cleared on $BITMAP.

But, scanning directory tree will use ntfs_index_next(),
that does not check $BITMAP.

So if index entries in $INDEX_ALLOCATION which is matched
with cleared $BITMAP is corrupted, ntfsck cannot find it.

Add checking entry's vcn if it is set or not in $BITMAP.
If not, it does not matched, so it initialize index entry.

Signed-off-by: JaeHoon Sim <jay.sim@lge.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>